feat: npx @nestjs/cli g app reservation
