+++ apps/reservation/src/main.ts
@@ -1,0 +1,8 @@
+import { NestFactory } from '@nestjs/core';
+import { ReservationModule } from './reservation.module';
+
+async function bootstrap() {
+  const app = await NestFactory.create(ReservationModule);
+  await app.listen(3000);
+}
+bootstrap();
