+++ apps/reservation/src/reservation.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class AppService {
+export class ReservationService {
   getHello(): string {
     return 'Hello World!';
   }
