+++ apps/reservation/test/app.e2e-spec.ts
@@ -1,14 +1,14 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
-import { AppModule } from './../src/app.module';
+import { ReservationModule } from './../src/reservation.module';
 
-describe('AppController (e2e)', () => {
+describe('ReservationController (e2e)', () => {
   let app: INestApplication;
 
   beforeEach(async () => {
     const moduleFixture: TestingModule = await Test.createTestingModule({
-      imports: [AppModule],
+      imports: [ReservationModule],
     }).compile();
 
     app = moduleFixture.createNestApplication();
